Cleanup/Improve CardinalModuleWidget hackery
